requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cssselect==1.1.0
Deprecated==1.2.13
fake-headers==1.0.2
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
html5lib==1.1
idna==3.3
loguru==0.6.0
lxml==4.9.1
multidict==6.0.2
packaging==21.3
pydantic==1.9.1
pyparsing==3.0.9
pyquery==1.4.3
PyYAML==6.0
redis==4.3.4
requests==2.28.1
retrying==1.3.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
webencodings==0.5.1
wrapt==1.14.1
yarl==1.7.2
